// src/app/cache-prefix.enum.ts
export enum CachePrefix {
  location = 'location',
  locationSection = 'locationSection',
  presentation = 'presentation',
  presentationSection = 'presentationSection',
  presentationSectionItem = 'presentationSectionItem',
  event = 'event',
  eventList = 'eventList',
  eventTicketAvailabilities = 'eventTicketAvailabilities',
}